# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Tutor API",
    "description": "API for AI Tutor, managing students, courses, and AI-generated questions. For more details, visit the [GitHub repository](https://github.com/ShishirShekhar/E-Learning-Platform.git).",
    "version": "1.0.0",
    "contact": {
      "name": "Shishir Shekhar",
      "email": "sspdav02@gmail.com"
    },
    "termOfService": "https://www.google.com",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
    "/api/v1/ai/generate-question": {
      "post": {
        "tags": [
          "AI"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Questions generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "questions": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "questions": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Topic is required to generate questions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Topic is required to generate questions"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Topic is required to generate questions"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Question generation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Question generation failed"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Question generation failed"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/related-documents": {
      "post": {
        "tags": [
          "AI"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Related documents fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "documents": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "documents": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Topic is required to fetch related documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Topic is required to fetch related documents"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Topic is required to fetch related documents"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch related documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch related documents"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch related documents"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/generate-summary": {
      "post": {
        "tags": [
          "AI"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Summary generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "summary": {
                          "type": "string",
                          "example": ""
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "summary": {
                          "type": "string",
                          "example": ""
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Topic is required to generate summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Topic is required to generate summary"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Topic is required to generate summary"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to generate summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Failed to generate summary"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Failed to generate summary"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/generate-content": {
      "post": {
        "tags": [
          "AI"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Content generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "content": {
                          "type": "string",
                          "example": ""
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "content": {
                          "type": "string",
                          "example": ""
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Topic is required to generate content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Topic is required to generate content"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Topic is required to generate content"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to generate content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Failed to generate content"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Failed to generate content"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}